services/CommunityService/requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
backoff==2.2.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
core @ file:./services/core
cryptography==43.0.3
Django==5.1.3
gql==3.5.0
graphene==3.4.3
graphene-directives==0.4.6
graphene-django==3.2.2
graphene-federation==3.2.0
graphene-file-upload==1.3.0
graphql-core==3.2.5
graphql-relay==3.2.0
idna==3.10
multidict==6.1.0
pillow==11.0.0
promise==2.3
propcache==0.2.0
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.0
